requirements.txt
flask
pandas
plotly
psycopg2-binary
pymysql
python-dotenv
cryptography